fix: review 1 반영
