+++ src/main/java/com/cholog/roomescape/RoomEscapeApplication.java
@@ -1,0 +1,10 @@
+package com.cholog.roomescape;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RoomEscapeApplication {
+    public static void main(String[] args) {
+
+    }
+}
