+++ src/main/java/com/cholog/roomescape/exception/GlobalExceptionHandler.java
@@ -1,0 +1,28 @@
+package com.cholog.roomescape.exception;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<Void> handleMethodArgumentNotValidException(
+            MethodArgumentNotValidException e
+    ) {
+        logger.warn(e.getMessage(), e);
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    public ResponseEntity<Void> handleRuntimeException(RuntimeException e){
+        logger.warn(e.getMessage(), e);
+        return ResponseEntity.notFound().build();
+    }
+}
