+++ src/main/java/com/cholog/roomescape/roomescape/exception/ReservationException.java
@@ -1,0 +1,7 @@
+package com.cholog.roomescape.roomescape.exception;
+
+public class ReservationException extends RuntimeException{
+    public ReservationException(String message) {
+        super(message);
+    }
+}
