+++ src/main/java/com/cholog/roomescape/roomescape/exception/ReservationNotFoundException.java
@@ -1,0 +1,9 @@
+package com.cholog.roomescape.roomescape.exception;
+
+import com.cholog.roomescape.roomescape.exception.code.RoomEscapeExceptionCode;
+
+public class ReservationNotFoundException extends RuntimeException {
+    public ReservationNotFoundException() {
+        super(RoomEscapeExceptionCode.RESERVATION_NOT_FOUND.getMessage());
+    }
+}
