+++ src/main/java/com/cholog/roomescape/roomescape/repository/ReservationRepository.java
@@ -1,0 +1,17 @@
+package com.cholog.roomescape.roomescape.repository;
+
+import com.cholog.roomescape.roomescape.entity.Reservation;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ReservationRepository {
+
+    Reservation save(Reservation reservation);
+
+    List<Reservation> findAll();
+
+    void delete(Reservation reservation);
+
+    Optional<Reservation> findById(Long reservationId);
+}
