--- src/main/java/exception/GlobalExceptionHandler.java
@@ -1,22 +1,0 @@
-package exception;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Void> handleRoomEscapeException(
-            MethodArgumentNotValidException e
-    ) {
-        logger.warn(e.getMessage(), e);
-        return ResponseEntity.badRequest().build();
-    }
-}
