--- src/main/java/roomescape/RoomEscapeApplication.java
@@ -1,11 +1,0 @@
-package roomescape;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RoomEscapeApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(RoomEscapeApplication.class, args);
-    }
-}
