--- src/main/java/roomescape/dto/response/ReservationResponse.java
@@ -1,23 +1,0 @@
-package roomescape.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.entity.Reservation;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record ReservationResponse(
-        Long id,
-        String name,
-
-        @JsonFormat(pattern = "yyyy-MM-dd")
-        LocalDate date,
-
-        @JsonFormat(pattern = "HH:mm")
-        LocalTime time
-) {
-
-    public static ReservationResponse toDto(Reservation reservation) {
-        return new  ReservationResponse(reservation.getId(), reservation.getName(), reservation.getDate(), reservation.getTime());
-    }
-}
