--- src/main/java/roomescape/entity/Reservation.java
@@ -1,66 +1,0 @@
-package roomescape.entity;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.Objects;
-
-public class Reservation {
-
-    private Long id;
-    private String name;
-    private LocalDate date;
-    private LocalTime time;
-
-    public Reservation() {
-    }
-
-    public Reservation(String name, LocalDate date, LocalTime time) {
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    private Reservation(Long id, String name, LocalDate date, LocalTime time) {
-        this.id = id;
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public LocalTime getTime() {
-        return time;
-    }
-
-    public static Reservation withId(Long id, Reservation reservation) {
-        return new Reservation(id, reservation.name, reservation.date, reservation.time);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Reservation other)) {
-            return false;
-        }
-        return id != null && id.equals(other.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-
-}
