--- src/main/java/roomescape/exception/ReservationNotFoundException.java
@@ -1,9 +1,0 @@
-package roomescape.exception;
-
-import roomescape.exception.code.RoomEscapeExceptionCode;
-
-public class ReservationNotFoundException extends RuntimeException {
-    public ReservationNotFoundException() {
-        super(RoomEscapeExceptionCode.RESERVATION_NOT_FOUND.getMessage());
-    }
-}
