--- src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,17 +1,0 @@
-package roomescape.repository;
-
-import roomescape.entity.Reservation;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ReservationRepository {
-
-    Reservation save(Reservation reservation);
-
-    List<Reservation> findAll();
-
-    void delete(Reservation reservation);
-
-    Optional<Reservation> findById(Long reservationId);
-}
