+++ src/main/resources/application.properties
@@ -1,7 +1,5 @@
 server.error.include-stacktrace=always
 
-spring.profiles.active=v2
-
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 spring.datasource.url=jdbc:h2:mem:database
